Moved SDF and V parser into the frontend (WIP linking)
