--- Code/Backend/Parser/ParsingTests/test.sdf
@@ -1,95 +1,0 @@
-(DELAYFILE
-    (SDFVERSION "2.1")
-    (DESIGN "FF1")
-    (VENDOR "verilog-to-routing")
-    (PROGRAM "vpr")
-    (VERSION "9.0.0-dev+v8.0.0-11943-g8cb20aa52-dirty")
-    (DIVIDER /)
-    (TIMESCALE 1 ps)
-
-    (CELL
-        (CELLTYPE "fpga_interconnect")
-        (INSTANCE routing_segment_D_output_0_0_to_lut_\$auto\$rtlil\.cc\:2714\:MuxGate\$139_input_0_3)
-        (DELAY
-            (ABSOLUTE
-                (IOPATH datain dataout (1693.56:1693.56:1693.56) (1693.56:1693.56:1693.56))
-            )
-        )
-    )
-
-    (CELL
-        (CELLTYPE "fpga_interconnect")
-        (INSTANCE routing_segment_clk_output_0_0_to_latch_Q_clock_0_0)
-        (DELAY
-            (ABSOLUTE
-                (IOPATH datain dataout (10:10:10) (10:10:10))
-            )
-        )
-    )
-
-    (CELL
-        (CELLTYPE "fpga_interconnect")
-        (INSTANCE routing_segment_async_reset_output_0_0_to_lut_\$auto\$rtlil\.cc\:2714\:MuxGate\$139_input_0_1)
-        (DELAY
-            (ABSOLUTE
-                (IOPATH datain dataout (463.327:463.327:463.327) (463.327:463.327:463.327))
-            )
-        )
-    )
-
-    (CELL
-        (CELLTYPE "fpga_interconnect")
-        (INSTANCE routing_segment_latch_Q_output_0_0_to_lut_\$auto\$rtlil\.cc\:2714\:MuxGate\$139_input_0_2)
-        (DELAY
-            (ABSOLUTE
-                (IOPATH datain dataout (0:0:0) (0:0:0))
-            )
-        )
-    )
-
-    (CELL
-        (CELLTYPE "fpga_interconnect")
-        (INSTANCE routing_segment_latch_Q_output_0_0_to_Q_input_0_0)
-        (DELAY
-            (ABSOLUTE
-                (IOPATH datain dataout (2304.7:2304.7:2304.7) (2304.7:2304.7:2304.7))
-            )
-        )
-    )
-
-    (CELL
-        (CELLTYPE "fpga_interconnect")
-        (INSTANCE routing_segment_lut_\$auto\$rtlil\.cc\:2714\:MuxGate\$139_output_0_0_to_latch_Q_input_0_0)
-        (DELAY
-            (ABSOLUTE
-                (IOPATH datain dataout (96:96:96) (96:96:96))
-            )
-        )
-    )
-
-    (CELL
-        (CELLTYPE "LUT_K")
-        (INSTANCE lut_\$auto\$rtlil\.cc\:2714\:MuxGate\$139)
-        (DELAY
-            (ABSOLUTE
-                (IOPATH in[1] out (152:152:152) (152:152:152))
-                (IOPATH in[2] out (150:150:150) (150:150:150))
-                (IOPATH in[3] out (150:150:150) (150:150:150))
-            )
-        )
-    )
-    
-    (CELL
-        (CELLTYPE "DFF")
-        (INSTANCE latch_Q)
-        (DELAY
-            (ABSOLUTE
-                (IOPATH (posedge clock) Q (303:303:303) (303:303:303))
-            )
-        )
-        (TIMINGCHECK
-            (SETUP D (posedge clock) (-46:-46:-46))
-        )
-    )
-    
-)
