--- Code/Backend/Parser/ParsingTests/test.v
@@ -1,83 +1,0 @@
-//Verilog generated by VPR 9.0.0-dev+v8.0.0-11943-g8cb20aa52-dirty from post-place-and-route implementation
-module FF1 (
-    input \D ,
-    input \clk ,
-    input \async_reset ,
-    output \Q 
-);
-
-    //Wires
-    wire \D_output_0_0 ;
-    wire \clk_output_0_0 ;
-    wire \async_reset_output_0_0 ;
-    wire \latch_Q_output_0_0 ;
-    wire \lut_$auto$rtlil.cc:2714:MuxGate$139_output_0_0 ;
-    wire \lut_$auto$rtlil.cc:2714:MuxGate$139_input_0_3 ;
-    wire \latch_Q_clock_0_0 ;
-    wire \lut_$auto$rtlil.cc:2714:MuxGate$139_input_0_1 ;
-    wire \lut_$auto$rtlil.cc:2714:MuxGate$139_input_0_2 ;
-    wire \Q_input_0_0 ;
-    wire \latch_Q_input_0_0 ;
-
-    //IO assignments
-    assign \Q  = \Q_input_0_0 ;
-    assign \D_output_0_0  = \D ;
-    assign \clk_output_0_0  = \clk ;
-    assign \async_reset_output_0_0  = \async_reset ;
-
-    //Interconnect
-    fpga_interconnect \routing_segment_D_output_0_0_to_lut_$auto$rtlil.cc:2714:MuxGate$139_input_0_3  (
-        .datain(\D_output_0_0 ),
-        .dataout(\lut_$auto$rtlil.cc:2714:MuxGate$139_input_0_3 )
-    );
-
-    fpga_interconnect \routing_segment_clk_output_0_0_to_latch_Q_clock_0_0  (
-        .datain(\clk_output_0_0 ),
-        .dataout(\latch_Q_clock_0_0 )
-    );
-
-    fpga_interconnect \routing_segment_async_reset_output_0_0_to_lut_$auto$rtlil.cc:2714:MuxGate$139_input_0_1  (
-        .datain(\async_reset_output_0_0 ),
-        .dataout(\lut_$auto$rtlil.cc:2714:MuxGate$139_input_0_1 )
-    );
-
-    fpga_interconnect \routing_segment_latch_Q_output_0_0_to_lut_$auto$rtlil.cc:2714:MuxGate$139_input_0_2  (
-        .datain(\latch_Q_output_0_0 ),
-        .dataout(\lut_$auto$rtlil.cc:2714:MuxGate$139_input_0_2 )
-    );
-
-    fpga_interconnect \routing_segment_latch_Q_output_0_0_to_Q_input_0_0  (
-        .datain(\latch_Q_output_0_0 ),
-        .dataout(\Q_input_0_0 )
-    );
-
-    fpga_interconnect \routing_segment_lut_$auto$rtlil.cc:2714:MuxGate$139_output_0_0_to_latch_Q_input_0_0  (
-        .datain(\lut_$auto$rtlil.cc:2714:MuxGate$139_output_0_0 ),
-        .dataout(\latch_Q_input_0_0 )
-    );
-
-
-    //Cell instances
-    LUT_K #(
-        .K(5),
-        .LUT_MASK(32'b00000000000000000101000101000000)
-    ) \lut_$auto$rtlil.cc:2714:MuxGate$139  (
-        .in({
-            1'b0,
-            \lut_$auto$rtlil.cc:2714:MuxGate$139_input_0_3 ,
-            \lut_$auto$rtlil.cc:2714:MuxGate$139_input_0_2 ,
-            \lut_$auto$rtlil.cc:2714:MuxGate$139_input_0_1 ,
-            1'b0
-         }),
-        .out(\lut_$auto$rtlil.cc:2714:MuxGate$139_output_0_0 )
-    );
-
-    DFF #(
-        .INITIAL_VALUE(1'b0)
-    ) \latch_Q  (
-        .D(\latch_Q_input_0_0 ), 
-        .Q(\latch_Q_output_0_0 ), 
-        .clock(\latch_Q_clock_0_0 )
-    );
-
-endmodule
